+++ includes/zuplus-functions.php
@@ -247,13 +247,20 @@ class ZU_PlusFunctions {
 		return $content;
 	}
 	
-	public function fix_content($content) {
+	public function fix_content($content, $add_p = true, $preserve_br = false) {
 		$replace_tags_from_to = array (
 			'<br />' => '',
 			"<br />\n" => '',
 		);
-		
-		$fixed = preg_replace('/^\s|\s$/', '', strtr(trim($content), $replace_tags_from_to));
+
+		$preserve_tags_from_to = array (
+			'<br />' => '[_br_]',
+			"<br />\n" => '[_br_]',
+		);
+
+		$fixed = preg_replace('/^\s|\s$/', '', strtr(trim($content), $preserve_br ? $preserve_tags_from_to : $replace_tags_from_to));
+		if($preserve_br) $fixed = str_replace('[_br_]', '<br />', trim($fixed));
+		if($add_p) $fixed = preg_replace(['#<p>\s*<br\s*/>#i', '#<br\s*/>\s*</p>#i'], ['<p>', '</p>'], sprintf('<p>%1$s</p>', $fixed));	// remove <br> right after <p> & right before </p>
 		return trim($fixed);
 	}